+++ tests/Tests/Tools/CliToolkitScripts/GenerateEnvConfigTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace MagicPush\CliToolkit\Tests\Tests\Tools\CliToolkitScripts;
 
-use MagicPush\CliToolkit\Tests\Tests\TestCaseAbstract;
 use MagicPush\CliToolkit\Tools\CliToolkit\ScriptClasses\Generate\EnvironmentConfigFile;
 
 use function PHPUnit\Framework\assertFileExists;
@@ -14,25 +13,18 @@ use function PHPUnit\Framework\assertNull;
 use function PHPUnit\Framework\assertSame;
 use function PHPUnit\Framework\assertTrue;
 
-class GenerateEnvConfigTest extends TestCaseAbstract {
-    private const string LAUNCHER_PATH = __DIR__ . '/' . '../../../../tools/cli-toolkit/launcher.php';
-
-    private const string GENERATED_DIRECTORY_PATH = __DIR__ . '/generated';
-
+final class GenerateEnvConfigTest extends CliToolkitScriptTestAbstract {
     /** The path should contain 2+ directories to test that directories are created recursively. */
     private const string CONFIG_PATH = self::GENERATED_DIRECTORY_PATH . '/env-config/parametizer.env.json';
 
 
-    private string $subcommandName;
+    private string $scriptName;
 
 
     protected function setUp(): void {
         parent::setUp();
 
-        require_once __DIR__ . '/' . '../../../../tools/cli-toolkit/init-autoloader.php';
-        $this->subcommandName = EnvironmentConfigFile::getScriptName();
-
-        static::removeDirectoryRecursively(self::GENERATED_DIRECTORY_PATH);
+        $this->scriptName = EnvironmentConfigFile::getScriptName();
     }
 
 
@@ -42,10 +34,10 @@ class GenerateEnvConfigTest extends TestCaseAbstract {
      * @see EnvironmentConfigFile::execute()
      */
     public function testFailToCreateDirectory(): void {
-        self::assertExecutionErrorOutput(
-            self::LAUNCHER_PATH,
+        static::assertExecutionErrorOutput(
+            static::LAUNCHER_PATH,
             "Unable to create a directory: '/asd'",
-            "{$this->subcommandName} /asd",
+            "{$this->scriptName} /asd",
         );
     }
 
@@ -55,10 +47,10 @@ class GenerateEnvConfigTest extends TestCaseAbstract {
      * @see EnvironmentConfigFile::execute()
      */
     public function testFailToWriteIntoFile(): void {
-        self::assertExecutionErrorOutput(
-            self::LAUNCHER_PATH,
+        static::assertExecutionErrorOutput(
+            static::LAUNCHER_PATH,
             sprintf("Unable to write into '/root/%s'", basename(self::CONFIG_PATH)),
-            "{$this->subcommandName} /root",
+            "{$this->scriptName} /root",
         );
     }
 
@@ -75,9 +67,9 @@ class GenerateEnvConfigTest extends TestCaseAbstract {
                 /* #2 */ self::CONFIG_PATH,
                 /* #3 */ PHP_EOL,
             ),
-            self::assertNoErrorsOutput(
-                self::LAUNCHER_PATH,
-                "{$this->subcommandName} " . dirname(self::CONFIG_PATH),
+            static::assertNoErrorsOutput(
+                static::LAUNCHER_PATH,
+                "{$this->scriptName} " . dirname(self::CONFIG_PATH),
             )
                 ->getStdOut(),
         ) ;
@@ -101,19 +93,19 @@ class GenerateEnvConfigTest extends TestCaseAbstract {
         assertNull(json_decode(file_get_contents(self::CONFIG_PATH), true));
 
         // Trying to overwrite the file with correct data... But it fails without `--force`.
-        self::assertExecutionErrorOutput(
-            self::LAUNCHER_PATH,
+        static::assertExecutionErrorOutput(
+            static::LAUNCHER_PATH,
             "File '" . realpath(self::CONFIG_PATH) . "' already exists.",
-            "{$this->subcommandName} " . dirname(self::CONFIG_PATH),
+            "{$this->scriptName} " . dirname(self::CONFIG_PATH),
         );
 
         // The file still exists:
         assertFileExists(self::CONFIG_PATH);
 
         // Now when using `--force` the script launch is successful:
-        self::assertNoErrorsOutput(
-            self::LAUNCHER_PATH,
-            "{$this->subcommandName} --force " . dirname(self::CONFIG_PATH),
+        static::assertNoErrorsOutput(
+            static::LAUNCHER_PATH,
+            "{$this->scriptName} --force " . dirname(self::CONFIG_PATH),
         );
 
         // ... And the generated file is overwritten: